# auth/src/main/java/com/bytebard/auth/controllers/AuthController.java
package com.bytebard.auth.controllers;

import com.bytebard.auth.service.AuthService;
import com.bytebard.auth.types.ChangePasswordRequest;
import com.bytebard.auth.types.LoginRequest;
import com.bytebard.auth.types.LoginResponse;
import com.bytebard.core.api.constants.Routes;
import com.bytebard.core.api.types.MvcApiReponse;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RequestMapping(Routes.AUTH)
@RestController
public class AuthController {
    private final AuthService authService;

    public AuthController(AuthService authService) {
        this.authService = authService;
    }

    @PostMapping(Routes.LOGIN)
    public ResponseEntity<MvcApiReponse<LoginResponse>> login(@RequestBody LoginRequest loginRequest) {
        var result = authService.login(loginRequest);
        var loginResponse = new MvcApiReponse<>(result, HttpStatus.OK, true);
        return ResponseEntity.ok().body(loginResponse);
    }

    @PostMapping(Routes.CHANGE_PASSWORD)
    public ResponseEntity<MvcApiReponse<Object>> changePassword(@RequestBody ChangePasswordRequest request) {
        authService.changePassword(request);
        var loginResponse = new MvcApiReponse<>(null, HttpStatus.OK, true);
        return ResponseEntity.ok().body(loginResponse);
    }
}